refactor(auth): delete file route.helpers.ts and use new ROUTS(API, APP) in project
