+++ src/features/auth/ui/RequireAuth.tsx
@@ -3,7 +3,7 @@ import { PropsWithChildren, useEffect } from 'react'
 import { useRouter } from 'next/navigation'
 import { useMeQuery } from '@/features/auth/api/authApi'
 import { Loading } from '@/shared/ui'
-import { ROUTES } from '@/shared/constant/routes'
+import { APP_ROUTES } from '@/shared/constant/app-routes'
 
 export function RequireAuth({ children }: PropsWithChildren) {
   const { data, isLoading, isError } = useMeQuery()
@@ -11,7 +11,7 @@ export function RequireAuth({ children }: PropsWithChildren) {
 
   useEffect(() => {
     if (isError) {
-      router.replace(ROUTES.AUTH.LOGIN)
+      router.replace(APP_ROUTES.AUTH.LOGIN)
     }
   }, [isError, router])
